GetEntityTypeIDsFromProperties -> EntityTypeManager namespace
This function is useful outside of PropertyGraph and it initializes an
EntityTypeManager. So a sensible home for it is as a static function in
the EntityTypeManager class.

One downside of this move is that we have to de-parallelize the function.
EntityTypeManager is in libsupport and libsupport doesn't have access to
the libgalois thread pool. This slows the function down considerably,
but since it is only supposed to be used to update old graphs the
slowdown is tolerable.
